Remove TakesValue restriction for HideEnv*
