+++ tests/help.rs
@@ -549,60 +549,6 @@ FLAGS:
     -V, --version
             Print version information";
 
-#[cfg(feature = "env")]
-static HIDE_ENV: &str = "ctest 0.1
-
-USAGE:
-    ctest [OPTIONS]
-
-FLAGS:
-    -h, --help       Print help information
-    -V, --version    Print version information
-
-OPTIONS:
-    -c, --cafe <FILE>    A coffeehouse, coffee shop, or café.";
-
-#[cfg(feature = "env")]
-static SHOW_ENV: &str = "ctest 0.1
-
-USAGE:
-    ctest [OPTIONS]
-
-FLAGS:
-    -h, --help       Print help information
-    -V, --version    Print version information
-
-OPTIONS:
-    -c, --cafe <FILE>    A coffeehouse, coffee shop, or café. [env: ENVVAR=MYVAL]";
-
-#[cfg(feature = "env")]
-static HIDE_ENV_VALS: &str = "ctest 0.1
-
-USAGE:
-    ctest [OPTIONS]
-
-FLAGS:
-    -h, --help       Print help information
-    -V, --version    Print version information
-
-OPTIONS:
-    -c, --cafe <FILE>    A coffeehouse, coffee shop, or café. [env: ENVVAR]
-    -p, --pos <VAL>      Some vals [possible values: fast, slow]";
-
-#[cfg(feature = "env")]
-static SHOW_ENV_VALS: &str = "ctest 0.1
-
-USAGE:
-    ctest [OPTIONS]
-
-FLAGS:
-    -h, --help       Print help information
-    -V, --version    Print version information
-
-OPTIONS:
-    -c, --cafe <FILE>    A coffeehouse, coffee shop, or café. [env: ENVVAR=MYVAL]
-    -p, --pos <VAL>      Some vals [possible values: fast, slow]";
-
 static CUSTOM_HELP_SECTION: &str = "blorp 1.4
 
 Will M.
@@ -1729,114 +1675,6 @@ fn issue_1052_require_delim_help() {
         app,
         "test --help",
         REQUIRE_DELIM_HELP,
-        false
-    ));
-}
-
-#[cfg(feature = "env")]
-#[test]
-fn hide_env() {
-    use std::env;
-
-    env::set_var("ENVVAR", "MYVAL");
-    let app = App::new("ctest").version("0.1").arg(
-        Arg::new("cafe")
-            .short('c')
-            .long("cafe")
-            .value_name("FILE")
-            .hide_env(true)
-            .env("ENVVAR")
-            .about("A coffeehouse, coffee shop, or café.")
-            .takes_value(true),
-    );
-    assert!(utils::compare_output(app, "ctest --help", HIDE_ENV, false));
-}
-
-#[cfg(feature = "env")]
-#[test]
-fn show_env() {
-    use std::env;
-
-    env::set_var("ENVVAR", "MYVAL");
-    let app = App::new("ctest").version("0.1").arg(
-        Arg::new("cafe")
-            .short('c')
-            .long("cafe")
-            .value_name("FILE")
-            .hide_env(false)
-            .env("ENVVAR")
-            .about("A coffeehouse, coffee shop, or café.")
-            .takes_value(true),
-    );
-    assert!(utils::compare_output(app, "ctest --help", SHOW_ENV, false));
-}
-
-#[cfg(feature = "env")]
-#[test]
-fn hide_env_vals() {
-    use std::env;
-
-    env::set_var("ENVVAR", "MYVAL");
-    let app = App::new("ctest")
-        .version("0.1")
-        .arg(
-            Arg::new("pos")
-                .short('p')
-                .long("pos")
-                .value_name("VAL")
-                .possible_values(&["fast", "slow"])
-                .about("Some vals")
-                .takes_value(true),
-        )
-        .arg(
-            Arg::new("cafe")
-                .short('c')
-                .long("cafe")
-                .value_name("FILE")
-                .hide_env_values(true)
-                .env("ENVVAR")
-                .about("A coffeehouse, coffee shop, or café.")
-                .takes_value(true),
-        );
-    assert!(utils::compare_output(
-        app,
-        "ctest --help",
-        HIDE_ENV_VALS,
-        false
-    ));
-}
-
-#[cfg(feature = "env")]
-#[test]
-fn show_env_vals() {
-    use std::env;
-
-    env::set_var("ENVVAR", "MYVAL");
-    let app = App::new("ctest")
-        .version("0.1")
-        .arg(
-            Arg::new("pos")
-                .short('p')
-                .long("pos")
-                .value_name("VAL")
-                .possible_values(&["fast", "slow"])
-                .about("Some vals")
-                .takes_value(true),
-        )
-        .arg(
-            Arg::new("cafe")
-                .short('c')
-                .long("cafe")
-                .value_name("FILE")
-                .hide_possible_values(true)
-                .env("ENVVAR")
-                .about("A coffeehouse, coffee shop, or café.")
-                .takes_value(true),
-        );
-    assert!(utils::compare_output(
-        app,
-        "ctest --help",
-        SHOW_ENV_VALS,
         false
     ));
 }
